Moved postgres gateway into index and updated references
