+++ src/gateways/index.ts
@@ -2,6 +2,7 @@ import CrmTokenGateway from './crmTokenGateway';
 import CrmGateway from './crmGateway';
 import MatPostgresGateway from './matPostgresGateway';
 import V1MatAPIGateway from './v1MatAPIGateway';
+import PostgresConnection, { PostgresOptions } from '../lib/postgresConnection';
 
 // if (!process.env.CRM_TOKEN_API_URL) {
 //   throw new Error('Error: CRM_TOKEN_API_URL environment variable missing');
@@ -25,7 +26,17 @@ export const crmGateway = new CrmGateway(
   crmTokenGateway
 );
 
-export const matPostgresGateway = new MatPostgresGateway();
+const options: PostgresOptions = {
+  user: `${process.env.DB_USER}`,
+  password: `${process.env.DB_PASSWORD}`,
+  host: `${process.env.DB_HOST}`,
+  port: process.env.DB_PORT ? parseInt(process.env.DB_PORT) : 5432,
+  database: `${process.env.DB_NAME}`,
+};
+const pgConn = new PostgresConnection(options);
+const connection = pgConn.getConnection();
+
+export const matPostgresGateway = new MatPostgresGateway(connection);
 
 export const v1MatAPIGateway: V1MatAPIGateway = new V1MatAPIGateway(
   `${process.env.V1_MAT_API_URL}`,
