+++ src/pages/api/tasks/[id]/close.ts
@@ -1,10 +1,12 @@
 import { NextApiRequest } from 'next';
 import CloseTask from '../../../../usecases/api/closeTask';
-import { v1MatAPIGateway } from '../../../../gateways';
-import MatPostgresGateway from '../../../../gateways/matPostgresGateway';
 import { getTokenPayloadFromRequest } from '../../../../usecases/api/getTokenPayload';
 import { ApiResponse } from '../../../../interfaces/apiResponses';
-import { crmGateway } from '../../../../gateways';
+import {
+  v1MatAPIGateway,
+  crmGateway,
+  matPostgresGateway,
+} from '../../../../gateways';
 
 export default async (
   req: NextApiRequest,
@@ -24,8 +26,6 @@ export default async (
   if (!process.env.V1_MAT_API_URL || !process.env.V1_MAT_API_TOKEN) {
     return res.status(500).end();
   }
-
-  const matPostgresGateway = new MatPostgresGateway();
 
   const closeTask = new CloseTask({
     crmGateway,
